biome adaptation
add changes based on biome
