+++ src/components/providers/ggx_container.tsx
@@ -1,5 +1,5 @@
 "use client";
-
+import type React from "react";
 import { useEffect, useState } from "react";
 
 import GGXWallet from "@/services/ggx";
